pyproject.toml
[tool.poetry]
name = "cota_parlamentar"
version = "0.1.0"
description = ""
authors = ["Rodrigo Torres <ratorres7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.1.2"
tqdm = "^4.49.0"
selenium = "^3.141.0"
requests = "^2.23.0"
python-anticaptcha = "^0.7.1"
python-dotenv = "^0.14.0"
selenium-wire = "^2.1.1"
python3-anticaptcha = "^1.6.2"
seaborn = "^0.11.0"
pymongo = "^3.11.0"
pylint = "^2.6.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
